pita_tool/pyproject.toml
[build-system]
requires= ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pita_dataset"
dynamic = ["version"]
requires-python = ">=3.11"
description = "Download a benchmark dataset for visible watermark detection"
readme = "README.md"
keywords = ["watermark", "dataset", "object-detection"]
dependencies = [
    "wget==3.2",
    "pycocotools==2.0.7",
    "opencv-python==4.9.0.80",
    "pillow==10.2.0"
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pyre-check",
    "fixit",
]
cli = [
    "click==8.1.7",
]



[project.urls]
Repository = "https://github.com/OrdinaryDev83/dnn-watermark"

[project.scripts]
spam-cli = "spam:main_cli"